Finalised the Gear Station
Added the full Upgrading Process to the Gear Station, added a few recipes; including a new way to obtain the Prismaridian Armour, now more annoying than ever. Prismarine Armour now has an ID to be upgraded, the other sets will probably follow. With the Mineral Analyser, the Tool Rod, and the Advanced Pick, there are now three main ways where we can greatly expand the content in the future.